pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "kafka-a2a"
version = "0.1.0"
description = "Kafka-based Agent-to-Agent (A2A) protocol implementation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "K-A2A Contributors" }]
keywords = ["a2a", "agents", "kafka", "fastapi", "langchain", "langgraph"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "aiokafka>=0.11.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
server = [
  "fastapi>=0.111.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30.0",
]
auth = [
  "pyjwt[crypto]>=2.10.0",
]
lang = [
  "langchain-core>=0.2.0",
  "langgraph>=0.2.0",
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.5.0",
]

[project.scripts]
ka2a = "kafka_a2a.cli:main"

[project.urls]
Repository = "https://example.invalid/k-a2a"

[tool.hatch.build.targets.wheel]
packages = ["src/kafka_a2a"]

[tool.ruff]
line-length = 100
